+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Service worker registration helpers in JavaScript client
 * Added `baseWebUrl` option in JavaScript client for configuring HTTP API URL separately from WebSocket URL
 * Web Push example application with client, service worker, and server
+* "Generate VAPID" operation in App model to create VAPID key pairs
 
 ### Changed